+++ compare-frontend/src/app/components/wizard/step/step.component.ts
@@ -1,5 +1,7 @@
 import { Component, Input } from '@angular/core';
-import { RiesgosStep, StepState } from 'src/app/state/state';
+import { Store } from '@ngrx/store';
+import { RiesgosState, RiesgosStep } from 'src/app/state/state';
+import * as AppActions from 'src/app/state/actions';
 
 @Component({
   selector: 'app-step',
@@ -7,5 +9,15 @@ import { RiesgosStep, StepState } from 'src/app/state/state';
   styleUrls: ['./step.component.css']
 })
 export class StepComponent {
+  
   @Input() step!: RiesgosStep;
+  public focusedStep$ = this.store.select(state => {
+    return state.riesgos.focusState.focusedStep;
+  });
+
+  constructor(private store: Store<{ riesgos: RiesgosState }>) {}
+
+  public focus() {
+    this.store.dispatch(AppActions.stepSelect({ stepId: this.step.step.id }));
+  }
 }